enhance margin

+++ src/modules/sales-cr/data-form.js
@@ -434,8 +434,11 @@ export class DataForm {
               console.log(salesVoidsItem.discountNominal);
               item.specialDiscount = salesVoidsItem.specialDiscount + "";
               console.log(salesVoidsItem.specialDiscount);
-              item.margin = parseInt(salesVoidsItem.margin);
+              // item.margin = parseInt(salesVoidsItem.margin);
+              // console.log(salesVoidsItem.margin);
+              item.margin = parseFloat(salesVoidsItem.margin);
               console.log(salesVoidsItem.margin);
+
               this.data.items.push(item);
               this.sumRow(item);
             }
@@ -591,11 +594,16 @@ console.log(this.error)
         ? parseInt(eventDiscountNominal.detail)
         : parseInt(item.discountNominal || 0)
       : parseInt(itemDetail.discountNominal);
+    // var margin = eventMargin
+    //   ? eventMargin.srcElement.value
+    //     ? parseInt(eventMargin.srcElement.value)
+    //     : parseInt(eventMargin.detail || 0)
+    //   : parseInt(itemDetail.margin);
     var margin = eventMargin
       ? eventMargin.srcElement.value
-        ? parseInt(eventMargin.srcElement.value)
-        : parseInt(eventMargin.detail || 0)
-      : parseInt(itemDetail.margin);
+        ? parseFloat(eventMargin.srcElement.value)
+        : parseFloat(eventMargin.detail || 0)
+      : parseFloat(itemDetail.margin);
     itemDetail.total = 0;
     console.log(itemDetail);
     if (parseInt(itemDetail.quantity) > 0) {
@@ -799,6 +807,7 @@ console.log(this.error)
   setDate() {
     this.data.date = new Date(this.data.datePicker);
   }
+  
   refreshPromo(indexItem) {
     var getPromoes = [];
     var storeId = this.data.storeId;
